# docs/specifications/protocol/http-api.openapi.yaml
openapi: 3.0.3
info:
  title: AWSM Coordination HTTP API
  version: "1"
  description: >-
    Canonical pre-release HTTPS adapter for the transport-independent AWSM synchronization
    protocol. Opaque payload bytes are never described semantically by this contract.
servers:
  - url: /
security:
  - accountBearer: []
paths:
...
  /api/vaults:
    get:
      operationId: listVaults
      parameters:
        - $ref: "#/components/parameters/ProtocolVersion"
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Zero or one Account-owned Vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultList"
        default:
          $ref: "#/components/responses/Outcome"
    post:
      operationId: attachVault
      parameters:
        - $ref: "#/components/parameters/ProtocolVersion"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachVaultRequest"
      responses:
        "201":
          description: Provisional Vault and current active Generation upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachVaultResponse"
        default:
          $ref: "#/components/responses/Outcome"
...
  /api/vaults/{vaultId}/complete:
    parameters:
      - $ref: "#/components/parameters/VaultId"
      - $ref: "#/components/parameters/ProtocolVersion"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      operationId: completeVault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteVaultRequest"
      responses:
        "200":
          description: Active Vault replica
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
        default:
          $ref: "#/components/responses/Outcome"
...
  /api/vaults/{vaultId}/commits:
    parameters:
      - $ref: "#/components/parameters/VaultId"
      - $ref: "#/components/parameters/ProtocolVersion"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      operationId: commitEventClosure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitRequest"
      responses:
        "200":
          description: Durable Event closure publication or already-active closure acknowledgement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitResult"
        default:
          $ref: "#/components/responses/Outcome"
...
components:
  securitySchemes:
    accountBearer:
      type: http
      scheme: bearer
  headers:
    ProtocolVersion:
      required: true
      schema:
        type: string
        enum: ["1"]
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
  parameters:
    ProtocolVersion:
      in: header
      name: Awsm-Protocol-Version
      required: true
      schema:
        type: string
        enum: ["1"]
    RequestId:
      in: header
      name: Awsm-Request-ID
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    VaultId:
      in: path
      name: vaultId
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
    Outcome:
      description: Stable protocol outcome
      headers:
        Awsm-Protocol-Version:
          $ref: "#/components/headers/ProtocolVersion"
        Awsm-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Outcome"
  schemas:
...
    Base64Url24:
      type: string
      pattern: "^[A-Za-z0-9_-]{32}$"
...
    Uuid:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    Counter:
      type: integer
      minimum: 0
      maximum: 9007199254740991
    PositiveByteLength:
      type: integer
      minimum: 1
      maximum: 9007199254740991
    Sha256:
      type: string
      pattern: "^[A-Za-z0-9_-]{43}$"
    ObjectType:
      type: string
      enum: [Event, BundleDescriptor, Artifact, VaultGeneration]
    ObjectMetadata:
      type: object
      additionalProperties: false
      required: [objectId, objectType, byteLength, sha256]
      properties:
        objectId:
          $ref: "#/components/schemas/Uuid"
        objectType:
          $ref: "#/components/schemas/ObjectType"
        byteLength:
          $ref: "#/components/schemas/PositiveByteLength"
        sha256:
          $ref: "#/components/schemas/Sha256"
...
    AttachVaultRequest:
      type: object
      additionalProperties: false
      required:
        [vaultId, generationId, generationNumber, accountSlot, generationObject]
      properties:
        vaultId:
          $ref: "#/components/schemas/Uuid"
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        accountSlot:
          $ref: "#/components/schemas/AccountSlot"
        generationObject:
          $ref: "#/components/schemas/ObjectMetadata"
    AttachVaultResponse:
      type: object
      additionalProperties: false
      required: [vault, upload, ticket]
      properties:
        vault:
          $ref: "#/components/schemas/Vault"
        upload:
          $ref: "#/components/schemas/Upload"
        ticket:
          $ref: "#/components/schemas/TransferTicket"
    CompleteVaultRequest:
      type: object
      additionalProperties: false
      required: [generationId]
      properties:
        generationId:
          $ref: "#/components/schemas/Uuid"
    Vault:
      type: object
      additionalProperties: false
      required:
        [
          vaultId,
          state,
          generationId,
          generationNumber,
          headCursor,
          accountSlot,
        ]
      properties:
        vaultId:
          $ref: "#/components/schemas/Uuid"
        accountSlot:
          $ref: "#/components/schemas/AccountSlot"
        state:
          type: string
          enum: [Provisional, Active]
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        predecessorGenerationId:
          $ref: "#/components/schemas/Uuid"
        headCursor:
          $ref: "#/components/schemas/Counter"
    VaultList:
      type: object
      additionalProperties: false
      required: [vaults]
      properties:
        vaults:
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/Vault"
    AccountSlot:
      type: object
      additionalProperties: false
      required:
        [version, slotId, vaultId, accountKeyId, algorithm, nonce, ciphertext]
      properties:
        version:
          type: integer
          enum: [1]
        slotId:
          $ref: "#/components/schemas/Uuid"
        vaultId:
          $ref: "#/components/schemas/Uuid"
        accountKeyId:
          $ref: "#/components/schemas/Uuid"
        algorithm:
          type: string
          enum: ["wrap:xchacha20poly1305:account:v1"]
        nonce:
          $ref: "#/components/schemas/Base64Url24"
        ciphertext:
          type: string
          minLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
...
    Upload:
      type: object
      additionalProperties: false
      required:
        [
          uploadId,
          objectId,
          state,
          partSizeBytes,
          partCount,
          receivedParts,
          expiresAt,
        ]
      properties:
        uploadId:
          $ref: "#/components/schemas/Uuid"
        objectId:
          $ref: "#/components/schemas/Uuid"
        state:
          type: string
          enum: [Open, Assembling, Completed, Expired, AlreadyDurable]
        partSizeBytes:
          $ref: "#/components/schemas/PositiveByteLength"
        partCount:
          type: integer
          minimum: 1
          maximum: 10000
        receivedParts:
          type: array
          uniqueItems: true
          items:
            type: integer
            minimum: 0
        expiresAt:
          type: string
          format: date-time
...
    TransferTicket:
      type: object
      additionalProperties: false
      required: [method, url, expiresAt, requiredHeaders]
      properties:
        method:
          type: string
          enum: [GET, PUT]
        url:
          type: string
        urlTemplate:
          type: string
        expiresAt:
          type: string
          format: date-time
        requiredHeaders:
          type: object
          additionalProperties:
            type: string
...
    CommitRequest:
      type: object
      additionalProperties: false
      required:
        [generationId, generationNumber, eventObjectId, dependencyObjectIds]
      properties:
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        eventObjectId:
          $ref: "#/components/schemas/Uuid"
        dependencyObjectIds:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Uuid"
    CommitResult:
      type: object
      additionalProperties: false
      required:
        [
          eventObjectId,
          generationId,
          generationNumber,
          dependencyObjectIds,
          cursor,
          durabilityAcknowledged,
        ]
      properties:
        eventObjectId:
          $ref: "#/components/schemas/Uuid"
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        dependencyObjectIds:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        cursor:
          $ref: "#/components/schemas/Counter"
        durabilityAcknowledged:
          type: boolean
          enum: [true]
...
    Outcome:
      type: object
      additionalProperties: false
      required: [outcome, retryable, requestId]
      properties:
        outcome:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_UNAVAILABLE
            - ACCOUNT_UNAVAILABLE
            - ACCOUNT_INPUT_INVALID
            - RECENT_AUTHENTICATION_REQUIRED
            - AUTHORIZATION_DENIED
            - PROTOCOL_VERSION_UNSUPPORTED
            - REQUEST_INVALID
            - IDEMPOTENCY_CONFLICT
            - VAULT_ID_UNAVAILABLE
            - VAULT_ACCOUNT_LIMIT_REACHED
            - VAULT_IDENTITY_MISMATCH
            - VAULT_NOT_FOUND
            - VAULT_NOT_READY
            - OBJECT_ID_CONFLICT
            - OBJECT_NOT_DURABLE
            - OBJECT_NOT_ACTIVE
            - OBJECT_CHECKSUM_MISMATCH
            - OBJECT_LENGTH_MISMATCH
            - UPLOAD_EXPIRED
            - UPLOAD_PART_CONFLICT
            - TRANSFER_TICKET_INVALID
            - DEPENDENCY_INVALID
            - VAULT_GENERATION_SUPERSEDED
            - VAULT_HEAD_CHANGED
            - GENERATION_CANDIDATE_CONFLICT
            - GENERATION_REACHABILITY_INVALID
            - RECOVERY_NOT_FOUND
            - RECOVERY_EXPIRED
            - PURGE_IN_PROGRESS
            - SERVER_BUSY
            - STORAGE_UNAVAILABLE
        retryable:
          type: boolean
        requestId:
          $ref: "#/components/schemas/Uuid"
        retryAfterSeconds:
          type: integer
          minimum: 0
        relatedObjectId:
          $ref: "#/components/schemas/Uuid"
        currentGenerationId:
          $ref: "#/components/schemas/Uuid"
        currentGenerationNumber:
          $ref: "#/components/schemas/Counter"
        headCursor:
          $ref: "#/components/schemas/Counter"
...